change width fix

+++ XGrid/Window controllers/GridLines window controller/GridLinesPreferences.h
@@ -2,6 +2,6 @@
 
 @interface GridLinesPreferences : NSWindowController
 
-@property (nonatomic, copy) void (^onSliderChange)(int);
+@property (nonatomic, copy) void (^onSliderChange)(uint8);
 
 @end
