+++ XGrid/Window controllers/GridLines window controller/View controller/GridLinesPreferencesViewController.h
@@ -1,5 +1,5 @@
 #import <Cocoa/Cocoa.h>
 
 @interface GridLinesPreferencesViewController : NSViewController
-@property (nonatomic, copy) void (^onSliderChange)(int);
+@property (nonatomic, copy) void (^onSliderChange)(uint8);
 @end
